+++ Tests/SwiftStdlibTests/IntExtensionsTests.swift
@@ -106,6 +106,14 @@ final class IntExtensionsTests: XCTestCase {
         XCTAssert(42.roundToNearest(0) == 42)
     }
 
+    func testDownTo() {
+        var latestValue = 0
+        5.down(to: -2) { currentValue in
+            latestValue = currentValue
+        }
+        assert(latestValue == -2)
+    }
+
     func testOperators() {
         XCTAssertEqual(5 ** 2, 25)
         XCTAssert((5 ± 2) == (3, 7) || (5 ± 2) == (7, 3))